server/requirements.txt
djangorestframework
pymysql
tensorflow
pillow
numpy